Revert "refactoring: get user profile"
This reverts commit 4bab008ac3af31b57e72ae49756226f5781846dc.

+++ src/main/java/fpt/project/bsmart/entity/dto/mentor/TeachInformationDTO.java
@@ -5,7 +5,7 @@ public class TeachInformationDTO {
     private int numberOfClass = 0;
     private int numberOfMember = 0;
 
-    private Double scoreFeedback = 0.0 ;
+    private Long scoreFeedback = 0L ;
 
     private int numberOfFeedBack = 0;
 
@@ -33,11 +33,11 @@ public class TeachInformationDTO {
         this.numberOfMember = numberOfMember;
     }
 
-    public Double getScoreFeedback() {
+    public Long getScoreFeedback() {
         return scoreFeedback;
     }
 
-    public void setScoreFeedback(Double scoreFeedback) {
+    public void setScoreFeedback(Long scoreFeedback) {
         this.scoreFeedback = scoreFeedback;
     }
